requirements.txt
backcall==0.2.0
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
matplotlib==3.4.1
networkx==2.5.1
numpy==1.20.2
pandas==1.2.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
progressbar2==3.53.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.5.6
pytz==2021.1
pywin32==300
pyzmq==22.0.3
scipy==1.6.2
seaborn==0.11.1
six==1.15.0
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
